+++ plugin/src/main/groovy/org/openbakery/signing/ProvisioningProfileIdReader.groovy
@@ -15,8 +15,12 @@
  */
 package org.openbakery.signing
 
+import org.apache.commons.lang.time.DateUtils
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
+
+import java.text.DateFormat
+import java.text.SimpleDateFormat
 
 
 class ProvisioningProfileIdReader {
@@ -24,18 +28,35 @@ class ProvisioningProfileIdReader {
 	private static Logger logger = LoggerFactory.getLogger(ProvisioningProfileIdReader.class)
 
 
-	def readProvisioningProfileUUID(def source) {
+	boolean checkExpired(String source) {
+		def matcher = source =~ "<key>ExpirationDate</key>\\s*\\n\\s*<date>(.*?)</date>"
+		if (matcher.find()) {
+			String[] dateParsePatterns = ["yyyy-MM-dd'T'HH:mm:ssX"];
+			Date date = DateUtils.parseDate(matcher[0][1], dateParsePatterns);
+			if (date.before(new Date())) {
+				DateFormat formatter = DateFormat.getDateTimeInstance(DateFormat.LONG, DateFormat.LONG, Locale.getDefault());
+				throw new IllegalArgumentException("The Provisioning Profile has expired on " + formatter.format(date) );
+			}
+		}
+	}
+
+
+	String readProvisioningProfileUUID(def source) {
 		logger.debug("destinationRoot: {}", source);
 		if (!(source instanceof File)) {
 			source = new File(source.toString())
 		}
 		if (source.exists()) {
 			def mobileprovisionContent = source.text
+			checkExpired(mobileprovisionContent);
+
 			def matcher = mobileprovisionContent =~ "<key>UUID</key>\\s*\\n\\s*<string>(.*?)</string>"
-			def uuid = matcher[0][1]
-			return uuid;
+			if (matcher.find()) {
+				return matcher[0][1];
+			}
 		}
 		return null;
 	}
+
 
 }