+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 Changes
 - Pods can no be reinstalled with --refresh-dependencies: Issue #90
+- If the provisioning profile has expired, the build fails early now, not after the build when signing: Issue #83
 
 Bugfixes
 - Fixed that the provisioning profile is set: Issue #97